PoC for component logic placed within a directive
