+++ projects/storefrontlib/cms-components/content/banner/banner.component.ts
@@ -4,83 +4,14 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { ChangeDetectionStrategy, Component, HostBinding } from '@angular/core';
-import {
-  CmsBannerComponent,
-  CmsService,
-  Image,
-  ImageGroup,
-  PageType,
-  SemanticPathService,
-} from '@spartacus/core';
-import { Observable } from 'rxjs';
-import { take, tap } from 'rxjs/operators';
-import { CmsComponentData } from '../../../cms-structure/page/model/cms-component-data';
+import {ChangeDetectionStrategy, Component} from '@angular/core';
+import {BannerComponent} from "./banner.directive";
 
 @Component({
   selector: 'cx-banner',
   templateUrl: './banner.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class BannerComponent {
-  routerLink: string | any[] | undefined;
+export class CxBannerComponent extends BannerComponent{
 
-  @HostBinding('class') styleClasses: string | undefined;
-
-  data$: Observable<CmsBannerComponent> = this.component.data$.pipe(
-    tap((data) => {
-      this.setRouterLink(data);
-      this.styleClasses = data.styleClasses;
-    })
-  );
-
-  constructor(
-    protected component: CmsComponentData<CmsBannerComponent>,
-    protected urlService: SemanticPathService,
-    protected cmsService: CmsService
-  ) {}
-
-  /**
-   * Returns `_blank` to force opening the link in a new window whenever the
-   * `data.external` flag is set to true.
-   */
-  getTarget(data: CmsBannerComponent): string | null {
-    return data.external === 'true' || data.external === true ? '_blank' : null;
-  }
-
-  protected setRouterLink(data: CmsBannerComponent): void {
-    if (data.urlLink) {
-      this.routerLink = data.urlLink;
-    } else if (data.contentPage) {
-      this.cmsService
-        .getPage({
-          id: data.contentPage,
-          type: PageType.CONTENT_PAGE,
-        })
-        .pipe(take(1))
-        .subscribe((page) => {
-          this.routerLink = page?.label;
-        });
-    } else if (data.product) {
-      this.routerLink = this.urlService.transform({
-        cxRoute: 'product',
-        params: { code: data.product },
-      });
-    } else if (data.category) {
-      this.routerLink = this.urlService.transform({
-        cxRoute: 'category',
-        params: { code: data.category },
-      });
-    }
-  }
-
-  getImage(data: CmsBannerComponent): Image | ImageGroup | undefined {
-    if (data.media) {
-      if ('url' in data.media) {
-        return data.media as Image;
-      } else {
-        return data.media as ImageGroup;
-      }
-    }
-  }
 }
