+++ projects/storefrontlib/cms-components/content/banner/banner.module.ts
@@ -4,17 +4,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
-import {
-  CmsConfig,
-  FeaturesConfigModule,
-  provideDefaultConfig,
-} from '@spartacus/core';
-import { GenericLinkModule } from '../../../shared/components/generic-link/generic-link.module';
-import { MediaModule } from '../../../shared/components/media/media.module';
-import { BannerComponent } from './banner.component';
+import {CommonModule} from '@angular/common';
+import {NgModule} from '@angular/core';
+import {RouterModule} from '@angular/router';
+import {CmsConfig, FeaturesConfigModule, provideDefaultConfig,} from '@spartacus/core';
+import {GenericLinkModule} from '../../../shared/components/generic-link/generic-link.module';
+import {MediaModule} from '../../../shared/components/media/media.module';
+import {CxBannerComponent} from './banner.component';
 
 @NgModule({
   imports: [
@@ -28,18 +24,19 @@ import { BannerComponent } from './banner.component';
     provideDefaultConfig(<CmsConfig>{
       cmsComponents: {
         SimpleResponsiveBannerComponent: {
-          component: BannerComponent,
+          component: CxBannerComponent,
         },
         BannerComponent: {
-          component: BannerComponent,
+          component: CxBannerComponent,
         },
         SimpleBannerComponent: {
-          component: BannerComponent,
+          component: CxBannerComponent,
         },
       },
     }),
   ],
-  declarations: [BannerComponent],
-  exports: [BannerComponent],
+  declarations: [CxBannerComponent],
+  exports: [CxBannerComponent],
 })
-export class BannerModule {}
+export class BannerModule {
+}
